Work in progress, NOT buildable. Commit made due to the need of switching computers.
Deleted Action business objects, WorldManager due to the switch to factory-strategy approch to process CHPP files.
Added new strategies, factory, interfaces, constants and other related objects.
Changed wage property type on SeniorPlayer from byte to int to reflect the correct data type.
Modified Player object from Players CHPP file since some properties datatypes were incorrect and updated references.

+++ Hyperar.HattrickUltimate.BusinessObjects/App/SeniorPlayer.cs
@@ -149,7 +149,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.App
         /// <summary>
         /// Gets or sets the Wage.
         /// </summary>
-        public byte Wage { get; set; }
+        public int Wage { get; set; }
 
         #endregion Public Properties
